REFACTOR:Moved api call to page.tsx, refactor HomeClient to only handle UI
